requirements.txt
# Business Flask App Requirements
Flask==2.2.3
Werkzeug==2.2.3
appwrite==4.0.0
requests==2.31.0
python-dotenv==1.0.0
qrcode==7.4.2
pillow==10.4.0
gunicorn==21.2.0
